backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
scipy==1.10.1
PyJWT==2.8.0
Werkzeug==2.3.0
python-dotenv==1.0.0
